WIN32 & MSVC preprocessor
While building under VS 2010(win7), I found that some preprocessor is not correct. I was getting 100+ errors for non-windows includes being used. The following fixed the issues.

_WIN32, instead of WIN32
MSVC does not work, but _MS_VER will give a value for the version of VC being used. It will return 0 if not VC.